requirements.txt
streamlit
langchain
langgraph
langchain-openai
langchain-community
python-dotenv
requests
psycopg2-binary